refactor(google_sheets): keep the formatting vocabulary out of the app-wide constants
Follow-up to the constants move: the tables now live in a new
common/format-constants.mjs rather than common/constants.mjs, which is left
exactly as it was on main.

constants.mjs is imported by google_sheets.app.mjs, so every action and source
in the app bundles it — adding to it made CI ask for a version bump on 47
components for what is a code-organization change. The four formatting actions
are the only consumers of this vocabulary, so a sibling file keeps the coupling
honest and the separation the review asked for intact: format-utils.mjs still
exports functions and nothing else.

BORDER_STYLE_OPTIONS still derives from constants.mjs's BORDER_STYLES (same set
minus NONE, which the NONE preset covers) via a read-only import.

Verified: same 71 checks pass, and evals 17-30 re-run against republished
copies — 14/14, no tool errors, so the new import resolves in the real bundle.
